+++ crates/cubecl-core/src/runtime_tests/mod.rs
@@ -11,7 +11,6 @@ pub mod different_rank;
 pub mod index;
 pub mod launch;
 pub mod line;
-pub mod memcpy_async;
 pub mod metadata;
 pub mod pipeline;
 pub mod plane;
@@ -86,7 +85,6 @@ macro_rules! testgen_float {
         cubecl_core::testgen_launch!();
         cubecl_core::testgen_line!();
         cubecl_core::testgen_pipeline!();
-        cubecl_core::testgen_memcpy_async!();
         cubecl_core::testgen_plane!();
         cubecl_core::testgen_sequence!();
         cubecl_core::testgen_slice!();
